beginning work on selecting excerpt
