+++ crmetronome.ui/.env
@@ -1,0 +1,9 @@
+REACT_APP_API_URL="https://localhost:44321"
+REACT_APP_API_KEY="AIzaSyAxf_WFdhoJBGsaOLZ8NR5W2kSRkyZ9ZIQ"
+REACT_APP_AUTH_DOMAIN="lll-emporium.firebaseapp.com"
+REACT_APP_PROJECT_ID="lll-emporium"
+REACT_APP_STORAGE_BUCKET="lll-emporium.appspot.com"
+REACT_APP_MESSAGING_SENDER_ID="1060625149116"
+REACT_APP_APP_ID="1:1060625149116:web:d2faa6047d87fd3c2274b9"
+REACT_APP_MEASUREMENT_ID="G-ZVCG02ZRXW"
+CHOKIDAR_USEPOLLING=true
