+++ crmetronome.ui/package.json
@@ -9,12 +9,14 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^14.3.0",
+    "axios": "^1.2.1",
     "bootstrap": "^5.2.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.4.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.1",
+    "react-select": "^5.7.0",
     "react-strap": "0.0.1",
     "sass": "^1.54.0",
     "web-vitals": "^2.1.4"
