+++ crmetronome.ui/src/components/SequenceSelector.js
@@ -1,0 +1,16 @@
+import React, {useEffect} from 'react';
+import { propTypes } from 'prop-types';
+
+const SequenceSelector = (sequence) => {
+
+  useEffect = () => {
+
+
+  }
+}
+
+SequenceSelector.propTypes = {
+  sequence: propTypes.object
+};
+
+export default SequenceSelector;
