+++ crmetronome.ui/src/helpers/apiKeys.js
@@ -1,0 +1,6 @@
+
+const crmConfig = {
+  baseUrl: process.env.REACT_APP_API_URL
+};
+
+export default crmConfig;
